+++ src/Compute/Compute/ChangeLog.md
@@ -21,6 +21,7 @@
 -->
 ## Upcoming Release
 * Added property `TimeCreated` to the Virtual Machine and Virtual Machine Scale Set models.
+* For `New-AzGalleryImageVersion`, update property mapping for `Encryption`. 
 
 ## Version 4.31.0
 * Added Trusted Launch Generic Breaking Change warning for `New-AzVM`, `New-AzDisk` and `New-AzVMSS` cmdlets.