+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,7 @@
 
 -->
 ## Upcoming Release
+* For `New-AzGalleryImageVersion`, update property mapping for `Encryption`. 
 
 ## Version 4.32.0
 * Added the `TimeCreated` property to the Virtual Machine and Virtual Machine Scale Set models.